# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Carve API
  description: |
    The Carve language-learning API. This document is the source of truth for
    every `/v1` endpoint; the TypeScript client in `apps/web/src/lib/api.ts`
    is regenerated from it and schemathesis runs property-based HTTP fuzz
    against it as part of the L4 testing layer.
  version: 0.3.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.carve.app
    description: Production
security: []  # default: no auth; protected operations explicitly set `security: [{ bearer: [] }]`
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    LanguageCode:
      type: string
      enum: [ja, zh-cn, zh-tw, zh, ko, en]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    TokenizeRequest:
      type: object
      required: [text, language]
      properties:
        text:     { type: string, maxLength: 50000 }
        language: { $ref: '#/components/schemas/LanguageCode' }
        mode:     { type: string, enum: [A, B, C], description: "Sudachi mode (JA only)" }
    TokenizeResponse:
      type: object
      required: [tokens]
      properties:
        tokens:
          type: array
          items: { type: object, additionalProperties: true }
...
paths:
...
  /v1/nlp/tokenize:
    post:
      tags: [nlp]
      operationId: tokenize
      summary: Tokenize text
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TokenizeRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenizeResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: Unsupported language, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: NLP service unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...